Split sidebar into desktop and mobile
