+++ pages/_app.tsx
@@ -1,9 +1,10 @@
 import '@/styles/globals.css';
 import { Container, MainContent } from '@/styles/layout.styles';
-import { Sidebar } from '@/components/Sidebar/Sidebar';
+import { DesktopSidebar } from '@/components/Sidebar/DesktopSidebar';
 import { SessionProvider } from 'next-auth/react';
 import { Media, MediaContextProvider } from '@/lib/media';
 import type { AppProps } from 'next/app';
+import MobileSidebar from '@/components/mobile/MobileSidebar';
 
 export default function App({
   Component,
@@ -24,16 +25,17 @@ export default function App({
 const Layout = ({ children }: { children: React.ReactNode }) => {
   return (
     <MediaContextProvider disableDynamicMediaQueries>
-      {/* <Media lessThan="sm"> */}
-      {/* TODO: add layout for mobile */}
-      {/* <div>Hello mobile</div> */}
-      {/* </Media> */}
-      {/* <Media> */}
-      <Container>
-        <Sidebar />
+      <Media lessThan="sm">
+        {/* TODO: add layout for mobile */}
+        <MobileSidebar />
         <MainContent>{children}</MainContent>
-      </Container>
-      {/* </Media> */}
+      </Media>
+      <Media greaterThanOrEqual="sm">
+        <Container>
+          <DesktopSidebar />
+          <MainContent>{children}</MainContent>
+        </Container>
+      </Media>
     </MediaContextProvider>
   );
 };
